+++ CHANGES.rst
@@ -1,6 +1,9 @@
 Changes in 0.11.2 (2020-xx-xx)
 ===============================================
 
+Improvements:
+* Registration: Support email verification link from homeserver (#3167).
+
 Bug fix:
  * AuthenticationViewController: Remove fallback to matrix.org when authentication failed (PR #3165).